mod scan data

+++ js/protos/node/node_pb.js
@@ -6043,7 +6043,7 @@ proto.node.BtfsScanTab.deserializeBinaryFromReader = function(msg, reader) {
       msg.setNewOnlineMinersCount(value);
       break;
     case 19:
-      var value = /** @type {number} */ (reader.readUint32());
+      var value = /** @type {number} */ (reader.readInt64());
       msg.setMinersAllAmount(value);
       break;
     default:
@@ -6199,7 +6199,7 @@ proto.node.BtfsScanTab.serializeBinaryToWriter = function(message, writer) {
   }
   f = message.getMinersAllAmount();
   if (f !== 0) {
-    writer.writeUint32(
+    writer.writeInt64(
       19,
       f
     );
@@ -6578,7 +6578,7 @@ proto.node.BtfsScanTab.prototype.setNewOnlineMinersCount = function(value) {
 
 
 /**
- * optional uint32 miners_all_amount = 19;
+ * optional int64 miners_all_amount = 19;
  * @return {number}
  */
 proto.node.BtfsScanTab.prototype.getMinersAllAmount = function() {
